--- tests/containers/search/visualizations/rank/mocks/awardingAgencyHelper.js
@@ -1,15 +1,0 @@
-import { awardingAgency } from '../../mockVisualizations';
-
-// Spending by Category Search
-export const performSpendingByCategorySearch = () => (
-    {
-        promise: new Promise((resolve) => {
-            process.nextTick(() => {
-                resolve({
-                    data: awardingAgency
-                });
-            });
-        }),
-        cancel: jest.fn()
-    }
-);
