--- tests/containers/search/visualizations/rank/mocks/cfdaHelper.js
@@ -1,15 +1,0 @@
-import { cfda } from '../../mockVisualizations';
-
-// Spending by Category Search
-export const performSpendingByCategorySearch = () => (
-    {
-        promise: new Promise((resolve) => {
-            process.nextTick(() => {
-                resolve({
-                    data: cfda
-                });
-            });
-        }),
-        cancel: jest.fn()
-    }
-);
