--- tests/containers/search/visualizations/rank/mocks/fundingAgencyHelper.js
@@ -1,15 +1,0 @@
-import { fundingAgency } from '../../mockVisualizations';
-
-// Spending by Category Search
-export const performSpendingByCategorySearch = () => (
-    {
-        promise: new Promise((resolve) => {
-            process.nextTick(() => {
-                resolve({
-                    data: fundingAgency
-                });
-            });
-        }),
-        cancel: jest.fn()
-    }
-);
