--- tests/containers/search/visualizations/rank/mocks/industryCodeHelper.js
@@ -1,15 +1,0 @@
-import { industryCode } from '../../mockVisualizations';
-
-// Spending by Category Search
-export const performSpendingByCategorySearch = () => (
-    {
-        promise: new Promise((resolve) => {
-            process.nextTick(() => {
-                resolve({
-                    data: industryCode
-                });
-            });
-        }),
-        cancel: jest.fn()
-    }
-);
